translocation/requirements.txt
pymssql
python-dotenv
pylint
boto3